// uldb/ngx-unity/src/app/unity-services/network-summary/network-interface/network-interface.type.ts
export interface NetworkInterfaceDetails {
    host: Host;
    interface_name: string;
    interface_itemid: string;
    receive: Receive;
    transmit: Transmit;
    bandwidth: Bandwidth;
    speed: Speed;
    inbound_discarded: Inbound_discarded;
    inbound_with_error: Inbound_with_error;
    outbound_discarded: Outbound_discarded;
    outbound_with_error: Outbound_with_error;
}
export interface Host {
    name: string;
    host_id: number;
    device_uuid: string;
    device_type: string;
}
export interface Receive {
    value: string;
    graph_id: string;
}
export interface Transmit {
    name: string;
    value: string;
    item_id: string;
    graph_id: string;
}
export interface Bandwidth {
    name: string;
    value: string;
    item_id: string;
    converted_value: string;
    graph_id: string;
}
export interface Speed {
    name: string;
    value: string;
    item_id: string;
    converted_value: string;
    graph_id: string;
}
export interface Inbound_discarded {
    name: string;
    value: string;
    item_id: string;
    graph_id: string;
}
export interface Inbound_with_error {
    name: string;
    value: string;
    item_id: string;
    graph_id: string;
}
export interface Outbound_discarded {
    name: string;
    value: string;
    item_id: string;
    graph_id: string;
}
export interface Outbound_with_error {
    name: string;
    value: string;
    item_id: string;
    graph_id: string;
}